backend/requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# LLM & Embeddings
openai>=1.10.0
qdrant-client>=1.7.0
numpy>=1.24.0

# Search & Retrieval
rank-bm25>=0.2.2

# Russian NLP
pymorphy3>=2.0.0

# Logging & Utils
python-json-logger>=2.0.0
aiofiles>=23.2.0
sse-starlette>=2.0.0

# Database & Auth
supabase>=2.6.0
PyJWT>=2.10.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Benchmark
tqdm>=4.66.0
pandas>=2.0.0
